doc_ocr_manager/requirements.txt
streamlit==1.52.1
easyocr==1.7.2
opencv-python==4.10.0.84
Pillow==10.4.0
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.5
